Autocomplete: Isolate finding completer into a function.
